+++ oap-server/server-core/src/main/java/org/apache/skywalking/oap/server/core/analysis/manual/segment/SegmentRecord.java
@@ -22,12 +22,12 @@ import java.util.List;
 import lombok.Getter;
 import lombok.Setter;
 import org.apache.skywalking.oap.server.core.analysis.Stream;
-import org.apache.skywalking.oap.server.core.analysis.manual.searchtag.Tag;
 import org.apache.skywalking.oap.server.core.analysis.record.Record;
 import org.apache.skywalking.oap.server.core.analysis.worker.RecordStreamProcessor;
 import org.apache.skywalking.oap.server.core.source.DefaultScopeDefine;
 import org.apache.skywalking.oap.server.core.storage.annotation.BanyanDB;
 import org.apache.skywalking.oap.server.core.storage.annotation.Column;
+import org.apache.skywalking.oap.server.core.storage.annotation.SQLDatabase;
 import org.apache.skywalking.oap.server.core.storage.annotation.SuperDataset;
 import org.apache.skywalking.oap.server.core.storage.type.Convert2Entity;
 import org.apache.skywalking.oap.server.core.storage.type.Convert2Storage;
@@ -38,6 +38,7 @@ import org.apache.skywalking.oap.server.core.storage.type.StorageBuilder;
 public class SegmentRecord extends Record {
 
     public static final String INDEX_NAME = "segment";
+    public static final String ADDITIONAL_TABLE_TAG = "segment_tag";
     public static final String SEGMENT_ID = "segment_id";
     public static final String TRACE_ID = "trace_id";
     public static final String SERVICE_ID = "service_id";
@@ -92,14 +93,8 @@ public class SegmentRecord extends Record {
     @Setter
     @Getter
     @Column(columnName = TAGS, indexOnly = true)
+    @SQLDatabase.AdditionalEntity.OnlyAdditional(additionalTables = {ADDITIONAL_TABLE_TAG})
     private List<String> tags;
-    /**
-     * Tags raw data is a duplicate field of {@link #tags}. Some storage don't support array values in a single column.
-     * Then, those implementations could use this raw data to generate necessary data structures.
-     */
-    @Setter
-    @Getter
-    private List<Tag> tagsRawData;
 
     @Override
     public String id() {