+++ oap-server/server-storage-plugin/storage-jdbc-hikaricp-plugin/src/main/java/org/apache/skywalking/oap/server/storage/plugin/jdbc/h2/dao/H2LogQueryDAO.java
@@ -55,22 +55,17 @@ import static org.apache.skywalking.oap.server.core.analysis.manual.log.Abstract
 import static org.apache.skywalking.oap.server.core.analysis.manual.log.AbstractLogRecord.TIMESTAMP;
 import static org.apache.skywalking.oap.server.core.analysis.manual.log.AbstractLogRecord.TRACE_ID;
 import static org.apache.skywalking.oap.server.core.analysis.manual.log.AbstractLogRecord.TRACE_SEGMENT_ID;
+import static org.apache.skywalking.oap.server.storage.plugin.jdbc.h2.dao.H2TableInstaller.ID_COLUMN;
 
 public class H2LogQueryDAO implements ILogQueryDAO {
     private final JDBCHikariCPClient h2Client;
     private final ModuleManager manager;
-    private final int maxSizeOfArrayColumn;
-    private final int numOfSearchValuesPerTag;
     private List<String> searchableTagKeys;
 
     public H2LogQueryDAO(final JDBCHikariCPClient h2Client,
-                         final ModuleManager manager,
-                         final int maxSizeOfArrayColumn,
-                         final int numOfSearchValuesPerTag) {
+                         final ModuleManager manager) {
         this.h2Client = h2Client;
         this.manager = manager;
-        this.maxSizeOfArrayColumn = maxSizeOfArrayColumn;
-        this.numOfSearchValuesPerTag = numOfSearchValuesPerTag;
     }
 
     @Override
@@ -91,14 +86,20 @@ public class H2LogQueryDAO implements ILogQueryDAO {
                                                        .provider()
                                                        .getService(ConfigService.class);
             searchableTagKeys = Arrays.asList(configService.getSearchableLogsTags().split(Const.COMMA));
-            if (searchableTagKeys.size() > maxSizeOfArrayColumn) {
-                searchableTagKeys = searchableTagKeys.subList(0, maxSizeOfArrayColumn);
-            }
         }
         StringBuilder sql = new StringBuilder();
         List<Object> parameters = new ArrayList<>(10);
 
-        sql.append("from ").append(LogRecord.INDEX_NAME).append(" where ");
+        sql.append("from ").append(LogRecord.INDEX_NAME);
+        if (!CollectionUtils.isEmpty(tags)) {
+            for (int i = 0; i < tags.size(); i++) {
+                sql.append(" inner join ").append(AbstractLogRecord.ADDITIONAL_TABLE_TAG).append(" ");
+                sql.append(AbstractLogRecord.ADDITIONAL_TABLE_TAG + i);
+                sql.append(" on ").append(LogRecord.INDEX_NAME).append(".").append(ID_COLUMN).append(" = ");
+                sql.append(AbstractLogRecord.ADDITIONAL_TABLE_TAG + i).append(".").append(ID_COLUMN);
+            }
+        }
+        sql.append(" where ");
         sql.append(" 1=1 ");
         if (startSecondTB != 0 && endSecondTB != 0) {
             sql.append(" and ").append(AbstractLogRecord.TIME_BUCKET).append(" >= ?");
@@ -135,20 +136,14 @@ public class H2LogQueryDAO implements ILogQueryDAO {
         }
 
         if (CollectionUtils.isNotEmpty(tags)) {
-            for (final Tag tag : tags) {
-                final int foundIdx = searchableTagKeys.indexOf(tag.getKey());
+            for (int i = 0; i < tags.size(); i++) {
+                final int foundIdx = searchableTagKeys.indexOf(tags.get(i).getKey());
                 if (foundIdx > -1) {
-                    sql.append(" and (");
-                    for (int i = 0; i < numOfSearchValuesPerTag; i++) {
-                        final String physicalColumn = LogRecord.TAGS + "_" + (foundIdx * numOfSearchValuesPerTag + i);
-                        sql.append(physicalColumn).append(" = ? ");
-                        parameters.add(tag.toString());
-                        if (i != numOfSearchValuesPerTag - 1) {
-                            sql.append(" or ");
-                        }
-                    }
-                    sql.append(")");
+                    sql.append(" and ").append(AbstractLogRecord.ADDITIONAL_TABLE_TAG + i).append(".");
+                    sql.append(AbstractLogRecord.TAGS).append(" = ?");
+                    parameters.add(tags.get(i).toString());
                 } else {
+                    //If the tag is not searchable, but is required, then don't need to run the real query.
                     return new Logs();
                 }
             }