+++ oap-server/server-storage-plugin/storage-jdbc-hikaricp-plugin/src/main/java/org/apache/skywalking/oap/server/storage/plugin/jdbc/postgresql/PostgreSQLStorageProvider.java
@@ -116,9 +116,7 @@ public class PostgreSQLStorageProvider extends ModuleProvider {
         this.registerServiceImplementation(IBatchDAO.class, new H2BatchDAO(postgresqlClient, config.getMaxSizeOfBatchSql(), config.getAsyncBatchPersistentPoolSize()));
         this.registerServiceImplementation(
                 StorageDAO.class,
-                new H2StorageDAO(
-                        getManager(), postgresqlClient, config.getMaxSizeOfArrayColumn(), config.getNumOfSearchableValuesPerTag())
-        );
+                new H2StorageDAO(postgresqlClient));
         this.registerServiceImplementation(
                 INetworkAddressAliasDAO.class, new H2NetworkAddressAliasDAO(postgresqlClient));
 
@@ -126,35 +124,19 @@ public class PostgreSQLStorageProvider extends ModuleProvider {
         this.registerServiceImplementation(IMetricsQueryDAO.class, new PostgreSQLMetricsQueryDAO(postgresqlClient));
         this.registerServiceImplementation(
                 ITraceQueryDAO.class,
-                new PostgreSQLTraceQueryDAO(
-                        getManager(),
-                        postgresqlClient,
-                        config.getMaxSizeOfArrayColumn(),
-                        config.getNumOfSearchableValuesPerTag()
-                )
+                new PostgreSQLTraceQueryDAO(getManager(), postgresqlClient)
         );
         this.registerServiceImplementation(IBrowserLogQueryDAO.class, new PostgreSQLBrowserLogQueryDAO(postgresqlClient));
         this.registerServiceImplementation(
                 IMetadataQueryDAO.class, new H2MetadataQueryDAO(postgresqlClient, config.getMetadataQueryMaxSize()));
         this.registerServiceImplementation(IAggregationQueryDAO.class, new PostgreSQLAggregationQueryDAO(postgresqlClient));
-        this.registerServiceImplementation(IAlarmQueryDAO.class, new PostgreSQLAlarmQueryDAO(
-                postgresqlClient,
-                getManager(),
-                config.getMaxSizeOfArrayColumn(),
-                config.getNumOfSearchableValuesPerTag()
-        ));
+        this.registerServiceImplementation(IAlarmQueryDAO.class, new PostgreSQLAlarmQueryDAO(postgresqlClient, getManager()));
         this.registerServiceImplementation(
                 IHistoryDeleteDAO.class, new H2HistoryDeleteDAO(postgresqlClient));
         this.registerServiceImplementation(ITopNRecordsQueryDAO.class, new H2TopNRecordsQueryDAO(postgresqlClient));
         this.registerServiceImplementation(
                 ILogQueryDAO.class,
-                new PostgreSQLLogQueryDAO(
-                        postgresqlClient,
-                        getManager(),
-                        config.getMaxSizeOfArrayColumn(),
-                        config.getNumOfSearchableValuesPerTag()
-                )
-        );
+                new PostgreSQLLogQueryDAO(postgresqlClient, getManager()));
 
         this.registerServiceImplementation(IProfileTaskQueryDAO.class, new H2ProfileTaskQueryDAO(postgresqlClient));
         this.registerServiceImplementation(IProfileTaskLogQueryDAO.class, new H2ProfileTaskLogQueryDAO(postgresqlClient));
@@ -176,33 +158,11 @@ public class PostgreSQLStorageProvider extends ModuleProvider {
                 .provider()
                 .getService(ConfigService.class);
         final int numOfSearchableTags = configService.getSearchableTracesTags().split(Const.COMMA).length;
-        if (numOfSearchableTags * config.getNumOfSearchableValuesPerTag() > config.getMaxSizeOfArrayColumn()) {
-            throw new ModuleStartException("Size of searchableTracesTags[" + numOfSearchableTags
-                    + "] * numOfSearchableValuesPerTag[" + config.getNumOfSearchableValuesPerTag()
-                    + "] > maxSizeOfArrayColumn[" + config.getMaxSizeOfArrayColumn()
-                    + "]. Potential out of bound in the runtime.");
-        }
-        final int numOfSearchableLogsTags = configService.getSearchableLogsTags().split(Const.COMMA).length;
-        if (numOfSearchableLogsTags * config.getNumOfSearchableValuesPerTag() > config.getMaxSizeOfArrayColumn()) {
-            throw new ModuleStartException("Size of searchableLogsTags[" + numOfSearchableLogsTags
-                    + "] * numOfSearchableValuesPerTag[" + config.getNumOfSearchableValuesPerTag()
-                    + "] > maxSizeOfArrayColumn[" + config.getMaxSizeOfArrayColumn()
-                    + "]. Potential out of bound in the runtime.");
-        }
-        final int numOfSearchableAlarmTags = configService.getSearchableAlarmTags().split(Const.COMMA).length;
-        if (numOfSearchableAlarmTags * config.getNumOfSearchableValuesPerTag() > config.getMaxSizeOfArrayColumn()) {
-            throw new ModuleStartException("Size of searchableAlarmTags[" + numOfSearchableAlarmTags
-                    + "] * numOfSearchableValuesPerTag[" + config.getNumOfSearchableValuesPerTag()
-                    + "] > maxSizeOfArrayColumn[" + config.getMaxSizeOfArrayColumn()
-                    + "]. Potential out of bound in the runtime.");
-        }
 
         try {
             postgresqlClient.connect();
 
-            MySQLTableInstaller installer = new PostgreSQLTableInstaller(
-                    postgresqlClient, getManager(), config.getMaxSizeOfArrayColumn(), config.getNumOfSearchableValuesPerTag()
-            );
+            MySQLTableInstaller installer = new PostgreSQLTableInstaller(postgresqlClient, getManager());
             getManager().find(CoreModule.NAME).provider().getService(ModelCreator.class).addModelListener(installer);
         } catch (StorageException e) {
             throw new ModuleStartException(e.getMessage(), e);