+++ oap-server/server-storage-plugin/storage-jdbc-hikaricp-plugin/src/main/java/org/apache/skywalking/oap/server/storage/plugin/jdbc/postgresql/dao/PostgreSQLTraceQueryDAO.java
@@ -24,9 +24,8 @@ import org.apache.skywalking.oap.server.storage.plugin.jdbc.h2.dao.H2TraceQueryD
 
 public class PostgreSQLTraceQueryDAO extends H2TraceQueryDAO {
     
-    public PostgreSQLTraceQueryDAO(ModuleManager manager, JDBCHikariCPClient h2Client, int maxSizeOfArrayColumn,
-            int numOfSearchableValuesPerTag) {
-        super(manager, h2Client, maxSizeOfArrayColumn, numOfSearchableValuesPerTag);
+    public PostgreSQLTraceQueryDAO(ModuleManager manager, JDBCHikariCPClient h2Client) {
+        super(manager, h2Client);
     }
     
 }